Started using Cocos Particle Effects for explosions
Had a few spare minutes so played around with the Particle Effects system and was really impressed with the effects! I decided to completely ditch the old explosions class and just let the Game Scene take care of the explosions. I have tested it on 3GS and so far it still runs at 60 FPS with those fancy effects running through for each explosion as well as a nice "Galaxy" particle effect running full time on the player tank.

Only problem is that there are some memory leaks and I still haven't managed to track them down yet, will have to try debug that later.